Work on getting the formatted counter value based on format option (long, float, large)

+++ metricbeat/module/windows/perfmon/pdh_integration_windows_test.go
@@ -254,7 +254,7 @@ func TestLongOutputFormat(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	_, okLong := values[processorTimeCounter][0].Measurement.(int64)
+	_, okLong := values[processorTimeCounter][0].Measurement.(int32)
 
 	assert.True(t, okLong)
 }
